+++ src/components/sign/EmailAuth/index.tsx
@@ -4,8 +4,8 @@ import { ChangeEvent, useState } from 'react'
 import { ValidationMessage } from 'src/lib/constants/ValidationMessage'
 import getValidation from 'src/lib/util/getValidation'
 import { UserParam } from 'src/lib/types/UserInterface'
-import { emailAuthAPI } from 'src/lib/api/Sign/emailAuthAPI'
 import { useAppSelector } from 'src/lib/hooks'
+import { emailAuthAPI } from 'src/lib/api/auth/emailAuthAPI'
 import SignInput from '../signInput/SignInput'
 import * as S from './styles'
