+++ src/lib/api/auth/SignInAPI.ts
@@ -1,7 +1,7 @@
-import { AxiosResponse } from 'axios'
+import { AxiosError, AxiosResponse } from 'axios'
 import { SignInParam } from 'src/lib/types/UserInterface'
-import { SIGNIN_URL } from 'src/lib/constants/Url'
 import { setLocalStorageToken } from 'src/lib/util/localStorage'
+import { AUTH_URL } from 'src/lib/constants/Url'
 import HttpClient from '../httpClient'
 
 class SignInAPI extends HttpClient {
@@ -20,9 +20,9 @@ class SignInAPI extends HttpClient {
     return response
   }
 
-  protected _handleError = (error: any) => {
+  protected _handleError = (error: AxiosError) => {
     const { response: errorResponse } = error
-    const errorCode = errorResponse.status
+    const errorCode = errorResponse?.status
 
     switch (errorCode) {
       case 401:
@@ -32,11 +32,17 @@ class SignInAPI extends HttpClient {
         alert('요청하신 작업을 수행할 수 없습니다. 관리자에게 문의해주세요.')
         break
     }
-
     return errorCode
   }
 
-  public SignIn = (data: SignInParam) => this.instance.post(SIGNIN_URL, data)
+  public SignIn = (data: SignInParam) => this.instance.post(`${AUTH_URL}login/common`, data)
+
+  public gitHubLogin = (code: string | string[] | qs.ParsedQs | qs.ParsedQs[] | undefined) =>
+    this.instance.post(
+      `${AUTH_URL}/login/github?code=${code}`,
+      {},
+      { headers: { 'Content-Type': 'application/x-www-form-urlencoded' } },
+    )
 }
 
 const signInAPI = new SignInAPI()
