+++ src/lib/api/auth/emailAuthAPI.ts
@@ -1,4 +1,4 @@
-import { AxiosResponse } from 'axios'
+import { AxiosError, AxiosResponse } from 'axios'
 import { EMAILAUTH_URL } from '../../constants/Url'
 import HttpClient from '../httpClient'
 
@@ -13,7 +13,7 @@ class EmailAuth extends HttpClient {
     this.instance.interceptors.response.use(this._handleResponse, this._handleError)
   }
 
-  private _handleResponse = (response: AxiosResponse | any) => {
+  private _handleResponse = (response: AxiosResponse) => {
     const responseCode = response.status
     switch (responseCode) {
       case 200:
@@ -25,12 +25,12 @@ class EmailAuth extends HttpClient {
       default:
         break
     }
-    return responseCode
+    return response
   }
 
-  private _handleError = (error: any) => {
+  private _handleError = (error: AxiosError) => {
     const { response: errorResponse } = error
-    const errorCode = errorResponse.status
+    const errorCode = errorResponse?.status
 
     if (errorCode === 401) {
       alert('인증번호를 다시 확인해주세요. 5분이 경과한 경우 다시 이메일 인증을 진행해주세요.')
