+++ src/lib/constants/Url.ts
@@ -1,4 +1,5 @@
-export const SIGNIN_URL = '/api/auth/login/common'
+export const AUTH_URL = '/api/auth/'
+
 export const SIGNUP_URL = '/api/member'
 export const EMAILAUTH_URL = '/api/auth/email'
 export const EMAILAUTHCODE_URL = '/api/auth/email/confirm'
@@ -9,6 +10,5 @@ export const STUDY_URL = '/api/study'
 
 export const GITHUB_LOGIN_URL =
   'https://github.com/login/oauth/authorize?client_id=75cd6db13c045cad11e6&redirect_uri=http://localhost:3000/callback'
-export const GITHUB_AUTH_URL = '/api/auth/login/github'
 
 export const USER_URL = '/api/member/'
