+++ src/lib/store/signForm/signInFormSlice.ts
@@ -1,5 +1,5 @@
 import { createAsyncThunk, createSlice } from '@reduxjs/toolkit'
-import { signInAPI } from 'src/lib/api/Sign/SignInAPI'
+import { signInAPI } from 'src/lib/api/auth/SignInAPI'
 import { SignInParam } from '../../types/UserInterface'
 
 const initialSignInForm = {
